fixed error with duplicate address returned
Due to the migrations not originally making all objectId properties objectId. There was an error with the join.
The allowed for the lookup to show random results that were not expected which diverted our attention from the bug
Updated the migrations file to also make parish on the Address an ObjectId.
This means the stop gap measure we made in the schema to update the pipeline of the aggregate not work.
We had to also updated the lookup to remove the ObjectId from both the fields that should be ObjectId.